@@ -4,63 +4,64 @@
   "scripts": {
     "ng": "ng",
     "start": "ng serve --proxy-config proxy.conf.js",
-    "build": "ng build -prod --output-path=backend/src/main/resources/static",
+    "build": "ng build --prod --output-path=backend/src/main/resources/static",
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e"
   },
   "license": "ISC",
   "dependencies": {
-    "@angular/animations": "^5.2.10",
-    "@angular/cdk": "^5.2.4",
-    "@angular/common": "^5.2.10",
-    "@angular/compiler": "^5.2.10",
-    "@angular/core": "^5.2.10",
-    "@angular/flex-layout": "5.0.0-beta.14",
-    "@angular/forms": "^5.2.10",
-    "@angular/http": "^5.2.10",
-    "@angular/material": "^5.2.4",
-    "@angular/platform-browser": "^5.2.10",
-    "@angular/platform-browser-dynamic": "^5.2.10",
-    "@angular/router": "^5.2.10",
-    "apollo-angular": "^1.0.1",
-    "apollo-angular-link-http": "^1.0.2",
-    "apollo-cache-inmemory": "^1.1.12",
-    "apollo-client": "^2.2.8",
+    "@angular/animations": "^6.0.2",
+    "@angular/cdk": "^6.0.2",
+    "@angular/common": "^6.0.2",
+    "@angular/compiler": "^6.0.2",
+    "@angular/core": "^6.0.2",
+    "@angular/flex-layout": "6.0.0-beta.15",
+    "@angular/forms": "^6.0.2",
+    "@angular/http": "^6.0.2",
+    "@angular/material": "^6.0.2",
+    "@angular/platform-browser": "^6.0.2",
+    "@angular/platform-browser-dynamic": "^6.0.2",
+    "@angular/router": "^6.0.2",
+    "apollo-angular": "^1.1.0",
+    "apollo-angular-link-http": "^1.1.0",
+    "apollo-cache-inmemory": "^1.2.1",
+    "apollo-client": "^2.3.1",
     "apollo-link-context": "^1.0.8",
     "apollo-link-error": "^1.0.9",
     "chart.js": "^2.7.2",
-    "core-js": "^2.5.5",
+    "core-js": "^2.5.6",
     "graphql": "^0.13.2",
-    "graphql-tag": "^2.8.0",
+    "graphql-tag": "^2.9.2",
     "hammerjs": "^2.0.8",
     "moment": "^2.22.1",
-    "rxjs": "^5.5.10",
+    "rxjs": "^6.1.0",
     "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "^1.7.4",
-    "@angular/compiler-cli": "^5.2.10",
-    "@angular/language-service": "^5.2.10",
+    "@angular/cli": "^6.0.1",
+    "@angular/compiler-cli": "^6.0.2",
+    "@angular/language-service": "^6.0.2",
+    "@angular-devkit/build-angular": "~0.6.1",
     "@types/chartjs": "^0.0.31",
     "@types/hammerjs": "^2.0.0",
-    "@types/jasmine": "~2.5.53",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/node": "^6.0.106",
-    "codelyzer": "^4.2.1",
-    "jasmine-core": "~2.8.0",
+    "@types/jasmine": "~2.8.6",
+    "@types/jasminewd2": "~2.0.3",
+    "@types/node": "~8.9.4",
+    "codelyzer": "^4.3.0",
+    "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~2.0.0",
+    "karma": "^2.0.2",
     "karma-chrome-launcher": "~2.2.0",
     "karma-cli": "~1.0.1",
     "karma-coverage-istanbul-reporter": "^1.4.2",
-    "karma-jasmine": "^1.1.1",
+    "karma-jasmine": "^1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
     "karma-phantomjs-launcher": "^1.0.4",
-    "protractor": "~5.1.2",
-    "ts-node": "~4.1.0",
+    "protractor": "~5.3.0",
+    "ts-node": "~5.0.1",
     "tslint": "~5.9.1",
-    "typescript": "~2.6.2"
+    "typescript": "~2.7.2"
   },
   "repository": {}
 }
